+++ malachite-nz/src/test_util/generators/exhaustive.rs
@@ -101,7 +101,7 @@ use malachite_base::iterators::bit_distributor::BitDistributorOutputType;
 use malachite_base::iterators::iter_windows;
 use malachite_base::num::arithmetic::traits::{
     ArithmeticCheckedShl, CoprimeWith, DivRound, DivisibleBy, DivisibleByPowerOf2, EqMod,
-    EqModPowerOf2, Height, Parity, PowerOf2, Square,
+    EqModPowerOf2, Height, HeightRef, Parity, PowerOf2, Square,
 };
 use malachite_base::num::basic::floats::PrimitiveFloat;
 use malachite_base::num::basic::integers::PrimitiveInt;
@@ -333,6 +333,18 @@ pub fn exhaustive_natural_polynomial_natural_unsigned_triple_gen_var_1()
             pow += max(p.height_significant_bits(), x.significant_bits());
             (p, x, pow)
         }),
+    )
+}
+
+pub fn exhaustive_natural_polynomial_natural_natural_triple_gen_var_1()
+-> It<(NaturalPolynomial, Natural, Natural)> {
+    Box::new(
+        exhaustive_triples_xyy(exhaustive_natural_polynomials(), exhaustive_naturals()).map(
+            |(p, x, m)| {
+                let m = max(p.height_ref(), &x) + m + Natural::ONE;
+                (p, x, m)
+            },
+        ),
     )
 }