+++ malachite-nz/src/test_util/generators/random.rs
@@ -96,7 +96,8 @@ use malachite_base::foer_sequences::FoerSequence;
 use malachite_base::iterators::with_special_value;
 use malachite_base::num::arithmetic::traits::{
     ArithmeticCheckedShl, CeilingLogBase2, CoprimeWith, DivRound, DivisibleBy, DivisibleByPowerOf2,
-    EqMod, EqModPowerOf2, Height, Parity, PowerOf2, RoundToMultipleOfPowerOf2Assign, Square,
+    EqMod, EqModPowerOf2, Height, HeightRef, Parity, PowerOf2, RoundToMultipleOfPowerOf2Assign,
+    Square,
 };
 use malachite_base::num::basic::floats::PrimitiveFloat;
 use malachite_base::num::basic::integers::PrimitiveInt;
@@ -373,6 +374,36 @@ pub fn random_natural_polynomial_natural_unsigned_triple_gen_var_1(
         .map(|(p, x, mut pow)| {
             pow += max(p.height_significant_bits(), x.significant_bits());
             (p, x, pow)
+        }),
+    )
+}
+
+pub fn random_natural_polynomial_natural_natural_triple_gen_var_1(
+    config: &GenConfig,
+) -> It<(NaturalPolynomial, Natural, Natural)> {
+    Box::new(
+        random_triples_xyy(
+            EXAMPLE_SEED,
+            &|seed| {
+                random_natural_polynomials(
+                    seed,
+                    config.get_or("mean_bits_n", 64),
+                    config.get_or("mean_bits_d", 1),
+                    config.get_or("mean_length_n", 4),
+                    config.get_or("mean_length_d", 1),
+                )
+            },
+            &|seed| {
+                random_naturals(
+                    seed,
+                    config.get_or("mean_bits_n", 64),
+                    config.get_or("mean_bits_d", 1),
+                )
+            },
+        )
+        .map(|(p, x, m)| {
+            let m = max(p.height_ref(), &x) + m + Natural::ONE;
+            (p, x, m)
         }),
     )
 }